image-processor/requirements.txt
aiokafka
Pillow
google-cloud-storage